finished refactoring of code into modules
