--- requirements.txt
@@ -1,18 +1,0 @@
-Flask
-Flask-SQLAlchemy
-psycopg2-binary
-pylint-flask
-flask_marshmallow
-flask-cors
-gevent
-gunicorn
-marshmallow-sqlalchemy
-python-dotenv
-email_validator
-flask-bcrypt
-pyjwt==1.4.2
-flask-session
-flask-mail
-pytest
-pytest-flask
-pytest-flask-sqlalchemy
